[Add]ログイン機能

+++ app/Http/Controllers/Tweet/DeleteController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers\Tweet;
 use App\Models\Tweet;
 use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use App\Services\TweetService;
+use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 
 class DeleteController extends Controller
 {
@@ -14,8 +16,11 @@ class DeleteController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function __invoke(Request $request)
+    public function __invoke(Request $request, TweetService $tweetService)
     {
+        if (!$tweetService->checkOwnTweet($request->user()->id, $request->id())) {
+            throw new AccessDeniedHttpException();
+        }
         $tweetId = (int) $request->route('tweetId');
         $tweet = Tweet::where('id', $tweetId)->firstOrFail();
         $tweet->delete();
